Add Role module and OrderlyPhone app
Add a new Role feature in the backend (RoleController, RoleService + interface, RoleRepository, RoleRequest/RoleResponse) implementing CRUD, validation, name normalization and DTO mapping with proper error responses. Update backend pom to include Spring Security and JJW T (jjwt-api/jjwt-impl/jjwt-jackson v0.11.5) and ensure validation starter is present. Add a new Android app scaffold 'OrderlyPhone' (Gradle setup, app module, resources, themes, icons, tests and IDE config). Also include small updates to OrderlyTablet project files and dependency versions.

+++ frontend/OrderlyTablet/app/src/main/java/com/example/orderlytablet/response/OrderResponse.kt
@@ -1,5 +1,6 @@
 package com.example.orderlytablet.response
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class OrderResponse(
@@ -7,7 +8,7 @@ data class OrderResponse(
     val dateTime: String,
     val state: String,
     val paymentMethod: String,
-    val total: Double,
+    val total: BigDecimal,
     val idEmployee: Long,
     val idClient: Long,
     val restTable: RestTableResponse
